+++ packages/api/src/pre-start/env-required.ts
@@ -1,0 +1,24 @@
+/** Here developers should specify required ENV variables */
+
+const requiredEnvVariables = [
+  'PORT',
+  'HOST',
+  'ADMINS',
+  'FRONTEND_URL',
+  'MONGO_DB',
+  'MONGO_HOST',
+  'MONGO_PORT',
+  'MONGO_USERNAME',
+  'MONGO_PASSWORD',
+  'COOKIE_DOMAIN',
+  'COOKIE_PATH',
+  'SECURE_COOKIE',
+  'JWT_SECRET',
+  'COOKIE_SECRET',
+  'JWT_ACCESS_EXP',
+  'JWT_REFRESH_EXP',
+  'SERVER_URL',
+  'NODE_ENV',
+];
+
+export { requiredEnvVariables };