+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Updated Activation page to look and work similar to the login page
+- Prevent error display on Period Details Page due to improperly calling useVerifyQuantifierPoolSize
+- Ensure default logo image loads properly
+- Ensure receiver is parsed when praise is submitted via discord Android app (workaround for Discord Android app issue)
+
+### Added
+- Check for required env variables on launch of api and discord-bot
+- Manage messages used by discord-bot /forward command via praise app settings page
+
 ## [0.4.0] - 2022-04-14
 
 ### Added